+++ src/components/TeacherComponents/SearchResultsItem/SearchResultsItem.tsx
@@ -15,9 +15,9 @@ import LessonMetadata from "@/components/SharedComponents/LessonMetadata";
 import TagPromotional from "@/components/SharedComponents/TagPromotional";
 import SearchDropdown from "@/components/TeacherComponents/SearchDropdown";
 import {
-  IntegratedLessonOverviewLinkProps,
-  IntegratedUnitOverviewLinkProps,
+  LessonOverviewLinkProps,
   resolveOakHref,
+  UnitOverviewLinkProps,
 } from "@/common-lib/urls";
 import { PathwaySchemaCamel } from "@/context/Search/search.types";
 import { NEW_COHORT } from "@/config/cohort";
@@ -46,11 +46,11 @@ export type SearchResultsItemProps = {
   | {
       type: "unit";
       nullTitle?: string;
-      buttonLinkProps: IntegratedUnitOverviewLinkProps;
+      buttonLinkProps: UnitOverviewLinkProps;
     }
   | {
       type: "lesson";
-      buttonLinkProps: IntegratedLessonOverviewLinkProps;
+      buttonLinkProps: LessonOverviewLinkProps;
       unitTitle: string;
     }
 );